Projekat2/app/requirements.txt
paho-mqtt==1.6.1
pyserial==3.5
python-dotenv==1.0.0
certifi==2023.7.22
influxdb-client==1.37.0
paho-mqtt==1.6.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
